requirements.txt
feedparser==6.0.11
requests==2.32.3
PyYAML==6.0.2
sib-api-v3-sdk==7.6.0
python-dotenv==1.0.1
